api/requirements.txt
Django>=2.2.3,<2.3.0
djangorestframework>=3.10.3,<3.11.0
psycopg2>=2.8.4,<2.9.0
django-filter>=2.2.0,<2.3.0
requests>=2.22.0,<2.23.0
django-cors-headers>=3.2.1,<3.3.0

flake8>=3.7.8,<3.8.0
python-slugify>=4.0.1,<4.1.0